+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog 
 
+## 0.0.1
+
+- Added custom theme
+
+- Imported assets
+
+- Added custom widgets:
+  - Tile
+
 ## 0.0.0
 
 - Started working on the app